+++ finance/grc.fraudDetection/src/main/resources/META-INF/resources/index.html
