+++ finance/grc.fraudDetection/src/main/resources/schema.tql
