+++ finance/grc.fraudDetection/src/test/resources/testBanks.gql
