+++ finance/grc.fraudDetection/src/test/resources/testMerchants.gql
