+++ finance/grc.fraudDetection/src/test/resources/testTransactions.gql
