--- finance/grc.fraudDetection_web/pom.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0"?>
-<project
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <artifactId>fraud_web</artifactId>
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>quarkus-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <goals>
-              <goal>build</goal>
-              <goal>generate-code</goal>
-              <goal>generate-code-tests</goal>
-            </goals>
-          </execution>
-        </executions>
-        <extensions>true</extensions>
-        <groupId>${quarkus.platform.group-id}</groupId>
-        <version>${quarkus.platform.version}</version>
-      </plugin>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <compilerArgs>
-            <arg>-parameters</arg>
-          </compilerArgs>
-        </configuration>
-        <version>${compiler-plugin.version}</version>
-      </plugin>
-      <plugin>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <configuration>
-          <systemPropertyVariables>
-            <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-            <maven.home>${maven.home}</maven.home>
-          </systemPropertyVariables>
-        </configuration>
-        <version>${surefire-plugin.version}</version>
-      </plugin>
-      <plugin>
-        <artifactId>maven-failsafe-plugin</artifactId>
-        <executions>
-          <execution>
-            <configuration>
-              <systemPropertyVariables>
-                <java.util.logging.manager>org.jboss.logmanager.LogManager
-                </java.util.logging.manager>
-                <maven.home>${maven.home}</maven.home>
-                <native.image.path>${project.build.directory}/${project.build.finalName}-runner
-                </native.image.path>
-              </systemPropertyVariables>
-            </configuration>
-            <goals>
-              <goal>integration-test</goal>
-              <goal>verify</goal>
-            </goals>
-          </execution>
-        </executions>
-        <version>${surefire-plugin.version}</version>
-      </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <artifactId>typedb-client</artifactId>
-      <groupId>com.vaticle.typedb</groupId>
-      <version>2.16.1</version>
-    </dependency>
-    <dependency>
-      <groupId>org.antlr</groupId>
-      <artifactId>antlr4-runtime</artifactId>
-      <version>4.8</version>
-    </dependency>
-    <dependency>
-      <groupId>com.opencsv</groupId>
-      <artifactId>opencsv</artifactId>
-      <version>5.7.1</version>
-    </dependency>
-    <dependency>
-      <artifactId>quarkus-smallrye-graphql</artifactId>
-      <groupId>io.quarkus</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>quarkus-arc</artifactId>
-      <groupId>io.quarkus</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>quarkus-junit5</artifactId>
-      <groupId>io.quarkus</groupId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <artifactId>${quarkus.platform.artifact-id}</artifactId>
-        <groupId>${quarkus.platform.group-id}</groupId>
-        <scope>import</scope>
-        <type>pom</type>
-        <version>${quarkus.platform.version}</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-  <groupId>com.vaticle</groupId>
-  <modelVersion>4.0.0</modelVersion>
-  <profiles>
-    <profile>
-      <activation>
-        <property>
-          <name>native</name>
-        </property>
-      </activation>
-      <id>native</id>
-      <properties>
-        <quarkus.package.type>native</quarkus.package.type>
-        <skipITs>false</skipITs>
-      </properties>
-    </profile>
-  </profiles>
-  <properties>
-    <compiler-plugin.version>3.10.1</compiler-plugin.version>
-    <maven.compiler.release>17</maven.compiler.release>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
-    <quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
-    <quarkus.platform.version>2.16.5.Final</quarkus.platform.version>
-    <skipITs>true</skipITs>
-    <surefire-plugin.version>3.0.0-M7</surefire-plugin.version>
-  </properties>
-  <repositories>
-    <repository>
-      <id>repo.vaticle.com</id>
-      <url>https://repo.vaticle.com/repository/maven/</url>
-    </repository>
-  </repositories>
-  <version>1.0-SNAPSHOT</version>
-</project>
