--- finance/grc.fraudDetection_web/src/main/java/com/typedb/examples/fraud/model/Bank.java
@@ -1,82 +1,0 @@
-/*
- * Copyright (C) 2022 Vaticle
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package com.typedb.examples.fraud.model;
-
-import java.util.Objects;
-
-public class Bank {
-
-  private String name;
-  private BankCoordinates coords;
-
-  public Bank() {}
-
-  public Bank(String name, BankCoordinates coords) {
-
-    this.name = name;
-    this.coords = coords;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public BankCoordinates getCoords() {
-    return coords;
-  }
-
-  public void setCoords(BankCoordinates coords) {
-    this.coords = coords;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof Bank bank)) {
-      return false;
-    }
-
-    return name.equals(bank.name) && coords.equals(bank.coords);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(name, coords);
-  }
-
-  @Override
-  public String toString() {
-
-    return
-        "bank {" +
-        "  name = '" + name + "', " +
-        "  coords = " + coords +
-        "}";
-  }
-}
